requirements.txt
streamlit
tensorflow==2.18.0
opencv-python-headless
Pillow
numpy==1.26.4
protobuf==3.20.3
gunicorn
